+++ appObjects/FlatCAMGerber.py
@@ -153,10 +153,10 @@ class GerberObject(FlatCAMObj, Gerber):
         self.calculations_finished.connect(self.update_area_chull)
 
         # Tools
-        self.ui.iso_button.clicked.connect(self.app.isolation_tool.run)
-        self.ui.generate_ncc_button.clicked.connect(self.app.ncclear_tool.run)
-        self.ui.generate_cutout_button.clicked.connect(self.app.cutout_tool.run)
-        self.ui.generate_follow_button.clicked.connect(self.app.follow_tool.run)
+        self.ui.iso_button.clicked.connect(lambda: self.app.isolation_tool.run(toggle=True))
+        self.ui.generate_ncc_button.clicked.connect(lambda: self.app.ncclear_tool.run(toggle=True))
+        self.ui.generate_cutout_button.clicked.connect(lambda: self.app.cutout_tool.run(toggle=True))
+        self.ui.generate_follow_button.clicked.connect(lambda: self.app.follow_tool.run(toggle=True))
 
         # Utilties
         self.ui.generate_bb_button.clicked.connect(self.on_generatebb_button_click)